create custom class BooksRecyclerView.kt, add it to layouts

+++ app/src/main/java/com/study/mybookshelf/model/Book.kt
@@ -4,7 +4,7 @@ open class Book (
     var title: String,
     var author: String,
     var photo: Int,
-    var raiting: Double,
+    var raiting: Float,
     var isDigital: Boolean,
     var comments: String) {
 }
