+++ app/src/main/java/com/study/mybookshelf/model/BorrowedBook.kt
@@ -6,7 +6,7 @@ data class BorrowedBook(
     var lbTitle: String,
     var lbAuthor: String,
     var lbPhoto: Int,
-    var lbRaiting: Double,
+    var lbRaiting: Float,
     var lbIsDigital: Boolean,
     var lbComments: String,
