+++ app/src/main/java/com/study/mybookshelf/ui/BorrowedBooksViewModel.kt
@@ -14,13 +14,13 @@ class BorrowedBooksViewModel : ViewModel() {
 
     init {
         val booksList = ArrayList<BorrowedBook>()
-        booksList.add(BorrowedBook("HarryPotter1", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter2", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter3", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter4", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter5", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter6", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
-        booksList.add(BorrowedBook("HarryPotter7", "J.K.Rowling", R.mipmap.ic_launcher, 5.0, true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter1", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter2", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter3", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter4", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter5", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter6", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
+        booksList.add(BorrowedBook("HarryPotter7", "J.K.Rowling", R.mipmap.ic_launcher, 5.0.toFloat(), true, "comment1", "Dasha", Date(2020, 12, 12), Date(2020, 12, 12)))
         borrowedBooksList.value = booksList
     }
 }
